+++ ao3kit/koreader/__init__.py
@@ -1,0 +1,23 @@
+"""KOReader collections deploy helpers."""
+
+from ao3kit.koreader.deploy import (
+    COLLECTIONS_JSON_NAME,
+    KOPLUGIN_DIRNAME,
+    build_collections_index,
+    deploy_metadata,
+    deploy_to_device,
+    install_plugin,
+    koreader_roots,
+    resolve_bundled_plugin_source,
+)
+
+__all__ = [
+    "COLLECTIONS_JSON_NAME",
+    "KOPLUGIN_DIRNAME",
+    "build_collections_index",
+    "deploy_metadata",
+    "deploy_to_device",
+    "install_plugin",
+    "koreader_roots",
+    "resolve_bundled_plugin_source",
+]